requirements.txt
requests==2.31.0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
